Install epel-release to meet aws-cfn-bootstrap dependency
The epel repo must be present both on the build host and in the
chroot environment. This patch updates both the build host dependencies
listed in the README, and the AWScliSetup.sh script that installs the
package to the chroot.

Defaulting to a url for EPELRELEASE was chosen for compatibility
across both CentOS and RHEL vendor repos, as the RHEL repos do not
include epel-release.

Fixes #33